Apply rustfmt.
Does not commit all changes rustfmt made, some changes, especially
to long strings, aren't good changes.

Most of the changes to long strings actually represent an underlying
problem in the test code, where many strings are repeated and the
testing is not handwritten, persay, because of the long (and somewhat
inconsistent, for unknown reasons) whitespace tokens.