Store token types that are of a subset in enum variants instead of using &str.
Keywords are a specific list of enum variants. In the future,
we should examine using the rust-phf crate to convert from the &str
values to the token type quickly (as that will benefit us in the
long run).

Storing keywords in a enum is also a good idea when considering the
parsing stage, where them being stored in an enum will allow for faster
comparison (e.g. when parsing a function, it should be faster to check
if the current token is "Token::Keyword(Keyword::Function)" than to
compare by testing two strings against each other. It should also be
more space-efficient to store keywords in an enum rather than as strings;
but this is an uncertainty at best.